+++ src/misc/BetterEditBugfixes.cpp
@@ -1,6 +1,7 @@
 #include <Geode/Geode.hpp>
 #include <Geode/modify/GJRotationControl.hpp>
 #include <alphalaneous.alphas_geode_utils/include/ObjectModify.hpp>
+#include "utils/Utils.hpp"
 
 using namespace geode::prelude;